+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## NEXT RELEASE
+
+* Moved all `lock` functionality to its own module
+* Added new `/locks/`, `/locks/{project_name}/enable`, and `/locks/{project_name}/disable` endpoints
+* Fixed a bug where retrieving a lock status may return the wrong value
+
 ## v0.17.0 (2022-02-12)
 
 * Overhauled the pipeline logs to provide more details that are easier to read